+++ src/dftbp/dftbplus/initprogram.F90
@@ -61,6 +61,7 @@ module dftbp_dftbplus_initprogram
   use dftbp_dftb_spin, only: Spin_getOrbitalEquiv, ud2qm, qm2ud
   use dftbp_dftb_thirdorder, only : TThirdOrderInp, TThirdOrder, ThirdOrder_init
   use dftbp_dftb_uniquehubbard, only : TUniqueHubbard, TUniqueHubbard_init
+  use dftbp_dftb_elecconstraints, only : TElecConstraint, TElecConstraint_init, TElecConstraintInput
   use dftbp_dftbplus_elstattypes, only : elstatTypes
   use dftbp_dftbplus_forcetypes, only : forceTypes
   use dftbp_dftbplus_inputdata, only : TParallelOpts, TInputData, TRangeSepInp, TControl, TBlacsOpts
@@ -750,6 +751,9 @@ module dftbp_dftbplus_initprogram
     !> Whether to run a range separated calculation
     logical :: isRangeSep
 
+    !> Whether constraints are imposed on electronic ground state
+    logical :: isElecconstr
+
     !> Range-separation data
     type(TRangeSepFunc), allocatable :: rangeSep
 
@@ -830,6 +834,9 @@ module dftbp_dftbplus_initprogram
 
     !> Write cavity information as cosmo file
     logical :: tWriteCosmoFile
+
+    !> Structure holding electronic constraints
+    type(TElecConstraint), allocatable :: elecConstrain
 
     !> Library interface handler
     type(TTBLite), allocatable :: tblite
@@ -968,6 +975,9 @@ module dftbp_dftbplus_initprogram
 
     !> Contains (iK, iS) tuples to be processed in parallel by various processor groups
     type(TParallelKS) :: parallelKS
+
+    !> True, if electron dynamics input block is present
+    logical :: isElecDyn
 
     !> Electron dynamics
     type(TElecDynamics), allocatable :: electronDynamics
@@ -1138,6 +1148,7 @@ module dftbp_dftbplus_initprogram
     procedure :: allocateDenseMatrices
     procedure :: getDenseDescCommon
     procedure :: ensureRangeSeparatedReqs
+    procedure :: ensureConstrainedDftbReqs
     procedure :: initRangeSeparated
     procedure :: initPlumed
 
@@ -1321,7 +1332,10 @@ contains
     this%tSpinOrbit = input%ctrl%tSpinOrbit
     this%tDualSpinOrbit = input%ctrl%tDualSpinOrbit
     this%t2Component = input%ctrl%t2Component
+    this%isXlbomd = allocated(input%ctrl%xlbomd)
     this%isRangeSep = allocated(input%ctrl%rangeSepInp)
+    this%isElecconstr = allocated(input%ctrl%elecConstrainInp)
+    this%isElecDyn = allocated(input%ctrl%elecDynInp)
 
     if (this%t2Component) then
       this%nSpin = 4
@@ -1470,7 +1484,7 @@ contains
       call error("SCC iterations must be larger than 0")
     end if
     if (this%tSccCalc) then
-      if (allocated(input%ctrl%elecDynInp)) then
+      if (this%isElecDyn) then
         if (input%ctrl%elecDynInp%tReadRestart .and. .not.input%ctrl%elecDynInp%tPopulations) then
           this%maxSccIter = 0
           this%isSccConvRequired = .false.
@@ -1862,7 +1876,7 @@ contains
      end if
 
      tRequireDerivator = this%tForces
-     if (.not. tRequireDerivator .and. allocated(input%ctrl%elecDynInp)) then
+     if (.not. tRequireDerivator .and. this%isElecDyn) then
        tRequireDerivator = input%ctrl%elecDynInp%tIons
      end if
      if (tRequireDerivator) then
@@ -1942,6 +1956,9 @@ contains
       end if
       if (this%isRangeSep) then
         call error("Range separated calculations do not yet work with transport calculations")
+      end if
+      if (this%isElecconstr) then
+        call error("Constrained DFTB calculations do not yet support electron transport.")
       end if
     end if
   #:endif
@@ -2226,6 +2243,12 @@ contains
       this%cutOff%mCutOff = max(this%cutOff%mCutOff, this%halogenXCorrection%getRCutOff())
     end if
 
+    if (allocated(input%ctrl%elecConstrainInp)) then
+      call this%ensureConstrainedDftbReqs(input%ctrl%elecConstrainInp)
+      allocate(this%elecConstrain)
+      call TElecConstraint_init(this%elecConstrain, input%ctrl%elecConstrainInp, this%orb)
+    end if
+
     this%tDipole = this%tMulliken
     if (this%tDipole) then
       block
@@ -2235,14 +2258,14 @@ contains
           call warning("Dipole printed for a charged system : origin dependent quantity")
           isDipoleDefined = .false.
         end if
-        if (this%tPeriodic.or.this%tHelical) then
+        if (this%tPeriodic .or. this%tHelical) then
           call warning("Dipole printed for extended system : value printed is not well defined")
           isDipoleDefined = .false.
         end if
         if (.not.isDipoleDefined) then
-          write(this%dipoleMessage, "(A)")"Warning: dipole moment is not defined absolutely!"
+          write(this%dipoleMessage, "(A)") "Warning: dipole moment is not defined absolutely!"
         else
-          write(this%dipoleMessage, "(A)")""
+          write(this%dipoleMessage, "(A)") ""
         end if
       end block
     end if
@@ -2563,7 +2586,6 @@ contains
     call this%initPlumed(env, input%ctrl%tPlumed, this%tMD, this%plumedCalc)
 
     ! Check for extended Born-Oppenheimer MD
-    this%isXlbomd = allocated(input%ctrl%xlbomd)
     if (this%isXlbomd) then
       if (input%ctrl%iThermostat /= 0) then
         call error("XLBOMD does not work with thermostats yet")
@@ -3556,7 +3578,7 @@ contains
     if (this%deltaDftb%isNonAufbau .and. allocated(this%ppRPA)) then
       call error("Delta DFTB incompatible with ppRPA")
     end if
-    if (this%deltaDftb%isNonAufbau .and. allocated(input%ctrl%elecDynInp)) then
+    if (this%deltaDftb%isNonAufbau .and. this%isElecDyn) then
       call error("Delta DFTB incompatible with electron dynamics")
     end if
     if (this%deltaDftb%isNonAufbau .and. this%tFixEf) then
@@ -3625,7 +3647,7 @@ contains
     end if
 
     ! Electron dynamics stuff
-    if (allocated(input%ctrl%elecDynInp)) then
+    if (this%isElecDyn) then
 
       if (this%t2Component) then
         call error("Electron dynamics is not compatibile with this spinor Hamiltonian")
@@ -5229,6 +5251,34 @@ contains
     end if
 
   end function getMinSccIters
+
+
+  !> Stop if any setting incompatible with the constrained DFTB formalism is found.
+  subroutine ensureConstrainedDftbReqs(this, elecConstrainInp)
+
+    !> Instance
+    class(TDftbPlusMain), intent(inout) :: this
+
+    !> Input parameters for electronic constraints
+    type(TElecConstraintInput), intent(in) :: elecConstrainInp
+
+    if (this%isXlbomd) then
+      call error("Constrained DFTB calculations do not yet support XLBOMD.")
+    end if
+
+    if (allocated(this%reks)) then
+      call error("Constrained DFTB calculations do not yet support REKS.")
+    end if
+
+    if (this%deltaDftb%isNonAufbau) then
+      call error("Constrained DFTB calculations do not yet support delta-DFTB.")
+    end if
+
+    if (this%isElecDyn) then
+      call error("Constrained DFTB calculations do not yet support electron dynamics.")
+    end if
+
+  end subroutine ensureConstrainedDftbReqs
 
 
   !> Stop if any range separated incompatible setting is found